Refactoring + clean up
* Fixed sharpmake compilation with latest Windows SDK on host
* Fixed warnings
* Refactored UI elements to easier to use interfaces
* Added frame speed independent camera movement with camera movement and rotation speed control with ImGui

+++ source/EngineCommon/System/InputManager.h
@@ -30,7 +30,7 @@ namespace engine::system
         SWITCH_CAMERA = 192,
     };
     constexpr auto KeyCount = magic_enum::enum_count<Key>();
-    inline u32 KeyToIndex(Key key) { return magic_enum::enum_index(key).value(); }
+    inline u32 KeyToIndex(Key key) { return (u32)magic_enum::enum_index(key).value(); }
 
     class KeyboardHandler 
     {
